clippy: merge src/tracker/tracker.rs into src/tracker/mod.rs
