+++ src/jobs/tracker_api.rs
@@ -4,7 +4,7 @@ use log::info;
 use tokio::task::JoinHandle;
 
 use crate::api::server;
-use crate::tracker::tracker::TorrentTracker;
+use crate::tracker::TorrentTracker;
 use crate::Configuration;
 
 pub fn start_job(config: &Configuration, tracker: Arc<TorrentTracker>) -> JoinHandle<()> {
